refactor: organize application code
