+++ apps/backend/src/application/dao/pet.dao.ts
@@ -1,0 +1,5 @@
+import { NearPet } from '../dto/near-pet.dto';
+
+export interface PetDao {
+  findNearest(): Promise<NearPet[]>;
+}
