+++ apps/backend/src/application/dto/near-pet.dto.ts
@@ -1,0 +1,3 @@
+export interface NearPet {
+  name: string;
+}
