+++ apps/backend/src/infra/controllers/pets.controller.ts
@@ -3,7 +3,7 @@ import { Body, Controller, Post, UseGuards } from '@nestjs/common';
 import { JwtPayload } from '../../auth/strategies/accessToken.strategy';
 import { User } from '../../decorators/user.decorator';
 import { AccessTokenGuard } from '../../guards/accessToken.guard';
-import PostPetForAdoption from '../../usecases/owner/post-pet-for-adoption/post-pet-for-adoption';
+import PostPetForAdoption from '../../application/usecases/owner/post-pet-for-adoption/post-pet-for-adoption';
 
 @Controller('api/v1/pets')
 export class PetsController {
