+++ apps/backend/src/modules/pet.module.ts
@@ -4,8 +4,8 @@ import { CaslModule } from 'nest-casl';
 import { PetsController } from '../infra/controllers/pets.controller';
 import { UserModule } from '../user/user.module';
 import { adoptionPermissions } from '../frameworks/casl/permissions/adoption.permissions';
-import PostPetForAdoption from '../usecases/owner/post-pet-for-adoption/post-pet-for-adoption';
-import GetPet from '../usecases/owner/get-pet/get-pet';
+import PostPetForAdoption from '../application/usecases/owner/post-pet-for-adoption/post-pet-for-adoption';
+import GetPet from '../application/usecases/owner/get-pet/get-pet';
 
 @Module({
   imports: [
